// src/app/generated_services/api2/model/personDetailResponse.ts
/**
 * API de Reconhecimento Facial com InsightFace
 * API para reconhecimento facial que suporta imagens com várias pessoas usando InsightFace
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FaceImageResponse } from './faceImageResponse';


/**
 * Schema for detailed Person response with images
 */
export interface PersonDetailResponse { 
    /**
     * Nome da pessoa
     */
    name: string;
    id: string;
    created_at: string;
    updated_at: string;
    is_active?: boolean;
    /**
     * Lista de imagens da pessoa
     */
    images?: Array<FaceImageResponse>;
}